code/requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.15.0
azure-search-documents>=11.7.0b2
openai>=1.40.0
python-dotenv
jupyter
ipykernel
